Add MAC command tests and improve MAC layer implementation (#184)
* Comment out logging in tests

* Refactor m_controlDataRate/DataRateAdaptation to ADR bit as in LoRaWAN specifications

* Add Margin unit of measure & getters for Margin and GwCnt

* Fit LinkAdrReq MAC command to specifications and remove regional dependencies

* Refactor MaxNumbTx to NbTrans as in LoRaWAN specifications + correct default value

* Avoid Object initialization for MAC commands (CreateObject -> Create)

* Rewrite OnLinkAdrReq from reputable source

* LoRaWAN Headers: store MAC commands to vector instead of list, clean-up and modernize c++, fix printing output

* Add getters to LinkAdrAns MAC command

* LinkCheckAns and LinkAdrReq tests

* DutyCycleReq: align terminology with specs, move logic from header to end-device, tests

* NS_TEST_ASSERT_MSG_EQ where needed

* Fix MAC command frequency encoding endianness

* RxParamSetupReq: add RxParamSetupAns getters, clean-up, fix frequency unit of measure, add tests

* DevStatusReq: implementation and tests

* Isolate unit test scope, add LinkAdrAns/RxParamSetupAns creation printing, clarify margin code

* Rework the SubBand class

- add missing EU868 radio sub-bands and fix tests
- downgrade SubBand from Object to SimpleRefCount
- remove unused default constructor/destructor
- refactor BelongsToSubBand to Contains for clarity
- re-add last frequency getter

* Rework the LogicalLoraChannel class

- downgrade LogicalLoraChannel from Object to SimpleRefCount
- remove unused default constructor/destructor and incomplete initializer constructor
- remove setters (constructor initialization only)
- refactor SetEnabledForUplink to EnableForUplink (align with DisableForUplink)

* LinkCheckAns: Fill margin field with meaningful values

* Rework LogicalLoraChannelHelper class

- redo the channel storage model to be aligned with LoRaWAN specifications: a fixed size array of channels accessed by index
- GetRawChannelArray replaces GetChannelList and GetEnabledChannelList as the only way to get channels: the MAC layer now performs its channel-related tasks (get wait time, get TX channel, change channel mask) outside of this class with just one sweep through of the array
- AddChannel becomes just SetChannel at index, in line with the channel storage structure found in the LoRaWAN specifications
- Similarly, discard RemoveChannel function: channels can be overwritten in the internal array with SetChannel as in specifications
- downgrade LogicalLoraChannelHelper from Object to SimpleRefCount
- add size initialization to constructor, remove default constructor
- remove aggregated duty-cycle management (already managed by external MAC layer)
- FIX duty-cycle next transmission time calculation in AddEvent
- add channel validity check
- add GetWaitingTime overload by frequency (that's what it is needed to get the wait time, no need to create an ad-hoc channel instance just to retrieve the wait time)
- add AddEvent overload by frequency (same reason as above)
- add GetTxPowerForChannel overload by frequency (same reason as above)
- remove channel addition by frequency, leading to incomplete tx channel creation (i.e. default values) + param duplication. This now mandates explicit external creation of channel instances to then be passed to the helper by pointer
- remove AddSubBand with init params, now only by pointer (same reason as above)

* Rework MacCommand classes

- remove unused destructors
- remove setters (mandate initializer constructor use)
- add missing getters for implemented commands
- add implementation TODOs
- improve printing and align teminology with specs
- add overflow asserts for field with size < 8 bits
- remove empty lines and boilerplate

* Rework EndDeviceLorawanMac class

- re-implement GetNextTransmissionDelay and GetChannelForTx
- remove now obsolete Shuffle method
- add setter and getter for transmission power
- correct ChMaskCntl spelling as per specifications
- remove AddSubBand and SetLogicalChannel
- move documentation to fix doxygen warning

* More LinkAdrReq test cases + use actual device matrix to test RX1Offset validity

* Implement OnNewChannelReq and add MAC command test cases

* Tx power should be converted to dBm ERP, not EIRP
+ more robust GetTxPowerIndex implementation

+++ model/adr-component.h
@@ -147,13 +147,12 @@ class AdrComponent : public NetworkControllerComponent
     double GetAverageSNR(EndDeviceStatus::ReceivedPacketList packetList, int historyRange);
 
     /**
-     * Get the LoRaWAN protocol TXPower configuration index from the Equivalent Isotropically
-     * Radiated Power (EIRP) in dBm.
+     * Get the LoRaWAN protocol TxPower parameter from the Equivalent Radiated Power (ERP) in dBm.
      *
-     * @param txPower Transission EIRP configuration.
-     * @return TXPower index as int.
+     * @param txPower Transission ERP configuration [dBm].
+     * @return TxPower parameter value as uint8_t.
      */
-    int GetTxPowerIndex(int txPower);
+    uint8_t GetTxPowerIndex(int txPower);
 
     enum CombiningMethod tpAveraging;      //!< TX power from gateways policy
     int historyRange;                      //!< Number of previous packets to consider
